requirements.txt
streamlit
opencv-python
pillow
bcrypt
razorpay
extra-streamlit-components